+++ src/module/attribution/AffiseModuleApi.ts
@@ -1,13 +1,7 @@
 import { AffiseModules } from "../AffiseModules";
 import type { OnKeyValueCallback } from "../OnKeyValueCallback";
 import type { AffiseModuleLinkApi} from "../link/AffiseModuleLinkApi"
-import type { AffiseLinkCallback} from "../link/AffiseLinkCallback"
-import type { AffiseProduct } from "../subscription/AffiseProduct";
 import type { AffiseModuleSubscriptionApi } from "../subscription/AffiseModuleSubscriptionApi";
-import type { AffiseProductType } from "../subscription/AffiseProductType";
-import type { AffisePurchasedInfo } from "../subscription/AffisePurchasedInfo";
-import type { AffiseProductsResult } from "../subscription/AffiseProductsResult";
-import type { AffiseResultCallback } from "../subscription/AffiseResultCallback";
 import type { AffiseModuleAppsFlyerApi } from "../appsflyer/AffiseModuleAppsFlyerApi";
 import type { AffiseModuleTikTokApi } from "../tiktok/AffiseModuleTikTokApi";
 import type { AffiseModuleAdvertisingApi } from "../advertising/AffiseModuleAdvertisingApi";
@@ -31,29 +25,4 @@ export interface AffiseModuleApi {
      * Get installed modules
      */
     getModulesInstalled(): Promise<AffiseModules[]>
-
-    /**
-     * @deprecated use Affise.module.link.resolve(url, callback)
-     * Module Link url Resolve
-     * @param url url
-     * @param callback redirected url
-     */
-    linkResolve(url: string, callback: AffiseLinkCallback): void;
-
-    /**
-     * @deprecated use Affise.module.subscription.fetchProducts(ids, callback)
-     * Module subscription fetchProducts
-     * @param ids list of product ids
-     * @param callback result callback
-     */
-    fetchProducts(ids: string[], callback: AffiseResultCallback<AffiseProductsResult>): void;
-
-    /**
-     * @deprecated use Affise.module.subscription.purchase(product, type, callback)
-     * Module subscription purchase
-     * @param product product
-     * @param type product type
-     * @param callback result callback
-     */
-    purchase(product: AffiseProduct, type: AffiseProductType, callback: AffiseResultCallback<AffisePurchasedInfo>): void;
 }